Update to version 2022
The simulator has been updated from 2021 to 2022. This commit introduce
the changes to have the code working for the new version of the
simulator

The new version only has dropdown menus, where the right choice is
always marked with a value that is different from "wrong". Some of them
have a fixed value, but one (the 6th drop down menu), has a changing
value for the right answer. For this reason a for loop has been employed
to detect the non-wrong option