Added exception control for corner cases
Some simulation pages seem to be different from most of those
experimented. In order to address this problem and avoid having the
bot terminated when an error occurs, this change has been introduced.

This commit introduces a bare try exept block on the main routine to
catch errors that may occurs and start a new simulation

When an error occurs, the simulation is lost and the bot waits for the
following available simulation